+++ catalyst/experiments/config.py
@@ -21,10 +21,10 @@ from catalyst.callbacks.optimizer import IOptimizerCallback, OptimizerCallback
 from catalyst.callbacks.scheduler import ISchedulerCallback, SchedulerCallback
 from catalyst.callbacks.timer import TimerCallback
 from catalyst.callbacks.validation import ValidationManagerCallback
+from catalyst.contrib.data.augmentor import Augmentor, AugmentorCompose
 from catalyst.core.callback import Callback
 from catalyst.core.experiment import IExperiment
 from catalyst.core.functional import check_callback_isinstance
-from catalyst.data.augmentor import Augmentor, AugmentorCompose
 from catalyst.registry import (
     CALLBACKS,
     CRITERIONS,
@@ -35,11 +35,9 @@ from catalyst.registry import (
 )
 from catalyst.typing import Criterion, Model, Optimizer, Scheduler
 from catalyst.utils.checkpoint import load_checkpoint, unpack_checkpoint
-from catalyst.utils.dict import merge_dicts
-from catalyst.utils.distributed import check_amp_available, get_rank
-from catalyst.utils.hash import get_short_hash
+from catalyst.utils.distributed import get_rank
 from catalyst.utils.loaders import get_loaders_from_params
-from catalyst.utils.misc import get_utcnow_time
+from catalyst.utils.misc import get_short_hash, get_utcnow_time, merge_dicts
 from catalyst.utils.torch import any2device, get_device, process_model_params