+++ catalyst/metrics/functional.py
@@ -387,11 +387,6 @@ def wrap_topk_metric2dict(
 
     Returns:
         wrapped metric function with List[Dict] output
-
-    Raises:
-        NotImplementedError: if metrics returned values are out of
-            torch.Tensor, int, float, Dict union.
-
     """
     metric_fn = partial(metric_fn, topk=topk_args)