+++ tests/_tests_scripts/dl_z_mvp_mnist_unet.py
@@ -7,8 +7,8 @@ from torch.nn import functional as F
 from torch.utils.data import DataLoader
 
 from catalyst import dl, metrics
+from catalyst.contrib.data.cv import ToTensor
 from catalyst.contrib.datasets import MNIST
-from catalyst.data.cv import ToTensor
 
 
 class ClassifyUnet(nn.Module):